more thorough benchmarks for Budu / Promise speed
