+++ benchmark/src/test/scala/io/chymyst/benchmark/ReactionDelaySpec.scala
@@ -141,48 +141,9 @@ class ReactionDelaySpec extends LogSpec {
 
     val offset = 250.0
     val power = -1.0
-    showRegression("reaction start delay", results.map(_._1 / 1000.0), x ⇒ math.pow(x + offset, power))
-    showRegression("emit delay", results.map(_._2 / 1000.0), x ⇒ math.pow(x + offset, power))
-    showRegression("reply delay", results.map(_._3 / 1000.0), x ⇒ math.pow(x + offset, power))
-  }
-
-  def showRegression(message: String, results: Seq[Double], funcX: Double => Double, funcY: Double => Double = identity): Unit = {
-    // Perform regression to determine the effect of JVM warm-up.
-    // Assume that the warm-up works as a0 + a1*x^(-c). Try linear regression with different values of c.
-    val dataX = results.indices.map(_.toDouble)
-    val dataY = results // pass with a min window
-      .zipAll(results.drop(1), Double.PositiveInfinity, Double.PositiveInfinity)
-      .zipAll(results.drop(2), (Double.PositiveInfinity, Double.PositiveInfinity), Double.PositiveInfinity)
-      .map { case ((x, y), z) ⇒ math.min(x, math.min(y, z)) }
-    val (a0, a1, a0stdev) = regressLSQ(dataX, dataY, funcX, funcY)
-    val speedup = f"${(a0 + a1 * funcX(dataX.head)) / (a0 + a1*funcX(dataX.last))}%1.2f"
-    println(s"Regression results for $message: constant = ${formatMicros(a0)} ± ${formatMicros(a0stdev)}, gain = ${formatMicros(a1)}*iteration, max. speedup = $speedup")
-
-    import org.sameersingh.scalaplot.Implicits._
-
-    val dataTheoryY = dataX.map(i ⇒ a0 + a1 * funcX(i))
-    val chart = xyChart(dataX → ((dataTheoryY, dataY)))
-    val plotter = new JFGraphPlotter(chart)
-    val plotdir = "logs/"
-    new File(plotdir).mkdir()
-    val plotfile = "benchmark " + message.replaceAll(" ", "_")
-    plotter.pdf(plotdir, plotfile)
-    println(s"Plot file produced in $plotdir$plotfile.pdf")
-  }
-
-  def det(a00: Double, a01: Double, a10: Double, a11: Double): Double = a00 * a11 - a01 * a10
-
-  def regressLSQ(xs: Seq[Double], ys: Seq[Double], funcX: Double ⇒ Double, funcY: Double ⇒ Double): (Double, Double, Double) = {
-    val n = xs.length
-    val sumX = xs.map(funcX).sum
-    val sumXX = xs.map(funcX).map(x ⇒ x * x).sum
-    val sumY = ys.map(funcY).sum
-    val sumXY = xs.zip(ys).map { case (x, y) ⇒ funcX(x) * funcY(y) }.sum
-    val detS = det(n.toDouble, sumX, sumX, sumXX)
-    val a0 = det(sumY, sumX, sumXY, sumXX) / detS
-    val a1 = det(n.toDouble, sumY, sumX, sumXY) / detS
-    val eps = math.sqrt(xs.zip(ys).map { case (x, y) ⇒ math.pow(a0 + a1 * funcX(x) - funcY(y), 2) }.sum) / n
-    (a0, a1, eps)
+    showRegression("reaction start delay", results.map(_._1.toDouble), x ⇒ math.pow(x + offset, power))
+    showRegression("emit delay", results.map(_._2.toDouble), x ⇒ math.pow(x + offset, power))
+    showRegression("reply delay", results.map(_._3.toDouble), x ⇒ math.pow(x + offset, power))
   }
 
   type Result = (Int, Int, Long, Boolean)
